llm-backend/requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
graphviz
requests
pydantic
gitpython
